tidyup mk includes

+++ bin/gh-includes/repo.sh
@@ -53,7 +53,7 @@ echo "INFO! - *eTagFile* : [ ${eTagFile} ]"
 #grep -oP '^ETag: \K["\w]+'
 
 if [ -e ${file} ] ;then
-   mv ${file} ${TEMP_DIR}/${fileName}
+   cp ${file} ${TEMP_DIR}/${fileName}
 fi
 
 if [ -e ${eTagFile} ] ;then
@@ -87,13 +87,13 @@ case "${doRequest}" in
     echo "OK! response ${doRequest}. OK"
     echo "INFO! response *stored* as: [ ${fileName} ]"
     echo "$(<${headerDump})" | grep -oP '^ETag: "\K(\w)+' > ${eTagFile}
-    rm ${TEMP_DIR}/${fileName}
+    [ -e ${TEMP_DIR}/${fileName} ] && rm ${TEMP_DIR}/${fileName}
     return 0
   ;;
   304)
     echo "OK! response ${doRequest}. OK"
     echo "INFO! ${fileName} is already up to date so will not be modified]"
-    mv ${TEMP_DIR}/${fileName} ${file}
+    [ -e ${TEMP_DIR}/${fileName} ] && mv ${TEMP_DIR}/${fileName} ${file}
     return 0
   ;;
   *)
